+++ docsource/releasenotes.rst
@@ -13,6 +13,7 @@ NEW FEATURES:
 
 * Added the ``getPermissionsMass`` administration command.
 * Added the ``getWorkspaceInfo`` administration command.
+* Added the ``listObjects`` administration command.
 
 UPDATED FEATURES / MAJOR BUG FIXES: